Move log_linear_ells to generators
